data-pipeline/requirements.txt
pyspark==3.5.0
psycopg2-binary
boto3
requests
apscheduler
python-dotenv